# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditResolutionSetResponse": {
        "properties": {
          "allowed_actions": {
            "items": {
              "$ref": "#/components/schemas/FrameworkAction"
            },
            "title": "Allowed Actions",
            "type": "array"
          },
          "framework_id": {
            "format": "uuid",
            "title": "Framework Id",
            "type": "string"
          },
          "knowledge_release_id": {
            "title": "Knowledge Release Id",
            "type": "string"
          },
          "resolution_set_id": {
            "format": "uuid",
            "title": "Resolution Set Id",
            "type": "string"
          },
          "resolutions": {
            "items": {
              "$ref": "#/components/schemas/AuditResolutionInput"
            },
            "title": "Resolutions",
            "type": "array"
          },
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          },
          "unresolved_blocking": {
            "title": "Unresolved Blocking",
            "type": "boolean"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "resolution_set_id",
          "framework_id",
          "revision_id",
          "version",
          "allowed_actions",
          "knowledge_release_id",
          "resolutions",
          "unresolved_blocking"
        ],
        "title": "AuditResolutionSetResponse",
        "type": "object"
      },
...
      "ConfirmFrameworkRequest": {
        "properties": {
          "audit_id": {
            "format": "uuid",
            "title": "Audit Id",
            "type": "string"
          },
          "expected_revision_id": {
            "format": "uuid",
            "title": "Expected Revision Id",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "resolution_set_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Set Id"
          }
        },
        "required": [
          "expected_revision_id",
          "expected_version",
          "audit_id"
        ],
        "title": "ConfirmFrameworkRequest",
        "type": "object"
      },
...
      "ConfirmedFrameworkResponse": {
        "properties": {
          "allowed_actions": {
            "items": {
              "$ref": "#/components/schemas/FrameworkAction"
            },
            "title": "Allowed Actions",
            "type": "array"
          },
          "audit": {
            "$ref": "#/components/schemas/FrameworkAuditResponse"
          },
          "confirmed_at": {
            "format": "date-time",
            "title": "Confirmed At",
            "type": "string"
          },
          "contract_version": {
            "title": "Contract Version",
            "type": "string"
          },
          "draft": {
            "$ref": "#/components/schemas/FrameworkDraftContract"
          },
          "framework_id": {
            "format": "uuid",
            "title": "Framework Id",
            "type": "string"
          },
          "knowledge_release_id": {
            "title": "Knowledge Release Id",
            "type": "string"
          },
          "resolutions": {
            "items": {
              "$ref": "#/components/schemas/AuditResolutionInput"
            },
            "title": "Resolutions",
            "type": "array"
          },
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          },
          "status": {
            "const": "confirmed",
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "framework_id",
          "task_id",
          "revision_id",
          "version",
          "status",
          "allowed_actions",
          "knowledge_release_id",
          "draft",
          "audit",
          "resolutions",
          "confirmed_at",
          "contract_version"
        ],
        "title": "ConfirmedFrameworkResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FrameworkResponse": {
        "properties": {
          "allowed_actions": {
            "items": {
              "$ref": "#/components/schemas/FrameworkAction"
            },
            "title": "Allowed Actions",
            "type": "array"
          },
          "draft": {
            "$ref": "#/components/schemas/FrameworkDraftContract"
          },
          "framework_id": {
            "format": "uuid",
            "title": "Framework Id",
            "type": "string"
          },
          "input": {
            "$ref": "#/components/schemas/FrameworkInputResponse"
          },
          "knowledge_release_id": {
            "title": "Knowledge Release Id",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/FrameworkStatus"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          },
          "unresolved_blocking_audit": {
            "title": "Unresolved Blocking Audit",
            "type": "boolean"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "framework_id",
          "task_id",
          "revision_id",
          "version",
          "status",
          "allowed_actions",
          "knowledge_release_id",
          "input",
          "draft",
          "unresolved_blocking_audit",
          "model"
        ],
        "title": "FrameworkResponse",
        "type": "object"
      },
...
      "RetryFrameworkReviewRequest": {
        "properties": {
          "expected_review_version": {
            "minimum": 1.0,
            "title": "Expected Review Version",
            "type": "integer"
          },
          "expected_revision_id": {
            "format": "uuid",
            "title": "Expected Revision Id",
            "type": "string"
          }
        },
        "required": [
          "expected_revision_id",
          "expected_review_version"
        ],
        "title": "RetryFrameworkReviewRequest",
        "type": "object"
      },
...
      "StartFrameworkReviewRequest": {
        "properties": {
          "expected_revision_id": {
            "format": "uuid",
            "title": "Expected Revision Id",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision_id",
          "expected_version"
        ],
        "title": "StartFrameworkReviewRequest",
        "type": "object"
      },
...
      "SubmitAuditResolutionsRequest": {
        "properties": {
          "audit_id": {
            "format": "uuid",
            "title": "Audit Id",
            "type": "string"
          },
          "expected_revision_id": {
            "format": "uuid",
            "title": "Expected Revision Id",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "resolutions": {
            "items": {
              "$ref": "#/components/schemas/AuditResolutionInput"
            },
            "minItems": 1,
            "title": "Resolutions",
            "type": "array"
          }
        },
        "required": [
          "expected_revision_id",
          "expected_version",
          "audit_id",
          "resolutions"
        ],
        "title": "SubmitAuditResolutionsRequest",
        "type": "object"
      },
...
      "UpdateFrameworkRequest": {
        "properties": {
          "draft": {
            "$ref": "#/components/schemas/FrameworkDraftContract"
          },
          "expected_revision_id": {
            "format": "uuid",
            "title": "Expected Revision Id",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "revision_reason": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Revision Reason",
            "type": "string"
          }
        },
        "required": [
          "expected_revision_id",
          "expected_version",
          "draft",
          "revision_reason"
        ],
        "title": "UpdateFrameworkRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "群学致知前后端架构基线 API。",
    "title": "群学致知 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/frameworks/{framework_id}": {
      "get": {
        "deprecated": true,
        "operationId": "get_framework",
        "parameters": [
          {
            "in": "path",
            "name": "framework_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Framework Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameworkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Framework",
        "tags": [
          "frameworks"
        ]
      },
      "patch": {
        "deprecated": true,
        "operationId": "update_framework",
        "parameters": [
          {
            "in": "path",
            "name": "framework_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Framework Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFrameworkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameworkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Update Framework",
        "tags": [
          "frameworks"
        ]
      }
    },
    "/api/frameworks/{framework_id}/audit-resolutions": {
      "post": {
        "deprecated": true,
        "operationId": "submit_audit_resolutions",
        "parameters": [
          {
            "in": "path",
            "name": "framework_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Framework Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAuditResolutionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditResolutionSetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Submit Audit Resolutions",
        "tags": [
          "frameworks"
        ]
      }
    },
    "/api/frameworks/{framework_id}/confirm": {
      "post": {
        "deprecated": true,
        "operationId": "confirm_framework",
        "parameters": [
          {
            "in": "path",
            "name": "framework_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Framework Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmFrameworkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmedFrameworkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Confirm Framework",
        "tags": [
          "frameworks"
        ]
      }
    },
    "/api/frameworks/{framework_id}/export": {
      "get": {
        "deprecated": true,
        "operationId": "export_confirmed_framework",
        "parameters": [
          {
            "in": "path",
            "name": "framework_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Framework Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormalFrameworkExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Export Confirmed Framework",
        "tags": [
          "frameworks"
        ]
      }
    },
    "/api/frameworks/{framework_id}/reviews": {
      "post": {
        "deprecated": true,
        "operationId": "start_framework_review",
        "parameters": [
          {
            "in": "path",
            "name": "framework_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Framework Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartFrameworkReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameworkReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Start Framework Review",
        "tags": [
          "frameworks"
        ]
      }
    },
    "/api/frameworks/{framework_id}/reviews/{review_run_id}": {
      "get": {
        "deprecated": true,
        "operationId": "get_framework_review",
        "parameters": [
          {
            "in": "path",
            "name": "framework_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Framework Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "review_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameworkReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Framework Review",
        "tags": [
          "frameworks"
        ]
      }
    },
    "/api/frameworks/{framework_id}/reviews/{review_run_id}/retry": {
      "post": {
        "deprecated": true,
        "operationId": "retry_framework_review",
        "parameters": [
          {
            "in": "path",
            "name": "framework_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Framework Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "review_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetryFrameworkReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameworkReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Retry Framework Review",
        "tags": [
          "frameworks"
        ]
      }
    },
...
    "/api/research-tasks/{task_id}/frameworks": {
      "post": {
        "deprecated": true,
        "operationId": "create_framework",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFrameworkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameworkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Create Framework",
        "tags": [
          "frameworks"
        ]
      }
    },
...
    "/api/session": {
      "get": {
        "operationId": "get_current_session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Get Current Session",
        "tags": [
          "session"
        ]
      }
    },
...
  }
}